+++ app/components/dashboard/dashboard-controller.js
@@ -3,8 +3,6 @@ angular.module('dashboard')
 
   this.state = State;
 
-  State.box.type = 'omnibox-dashboard';
-
   // dimensions are dependent on screen size.
   // this is calculated in directive.
   this.dims = {};
@@ -18,5 +16,4 @@ angular.module('dashboard')
 
   // default selection
   this.selectedStat = this.selectedStat || this.stats[2];
-
 }]);